Wait for index readiness in flaky RediSearch tests
RediSearch indexes documents asynchronously (keyspace-notification
driven for documents written after FT.CREATE), so queries issued right
after the writes intermittently observed a partially built index:
testWithSuffixTrieOption and testSvsVamanaWithDifferentDistanceMetrics
returned 0 results instead of the expected counts.

Extract the FT.INFO num_docs polling helper from
RediSearchIntegrationTests into a shared SearchTestSupport class
(rewritten on top of Wait with 10ms polls instead of 500ms sleeps) and
await index readiness before querying in both flaky tests.
