+++ src/test/java/io/lettuce/core/search/RediSearchVectorIntegrationTests.java
@@ -357,6 +357,9 @@ public class RediSearchVectorIntegrationTests {
             vec3.put(ByteBuffer.wrap("embedding".getBytes()), floatArrayToByteBuffer(new float[] { 0.5f, 0.5f, 0.0f }));
             redisBinary.hmset(ByteBuffer.wrap((metric.toLowerCase() + ":vec3").getBytes()), vec3);
 
+            // Let the notification-driven indexer catch up before querying
+            SearchTestSupport.awaitIndexReady(redis, indexName, 3);
+
             // Query with vector similar to vec1
             ByteBuffer queryVector = floatArrayToByteBuffer(new float[] { 0.9f, 0.1f, 0.0f });
             ByteBuffer blobKey = ByteBuffer.wrap("query_vec".getBytes());
